dev_requirements.txt
numpy==1.13.3
pillow